silence warnings during rspec test runs
If we're testing via rspec, we don't need to include any corosync
recipes, so we don't need to warn about anything.

However if we're *not* testing and we're still using Chef Solo, then
pacemaker::default can't possibly work, due to corosync::authkey
requiring Chef search.  In this case we need to abort.